Update index.ts
On Switch Only - change endpoint to "protection" and use configuration to supply time (in ms).
Currently works to turn off, but not back on.
Will require some if/else logic to use "duration" header based on whether turning on or off. Using enabled: true without duration: null results in failure.